email_microservice/requirements.txt
fastapi==0.115.6
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.2
uvicorn==0.34.0

